fixed image upload to gpu

+++ OptixPathtracer/source/ModelLoading/Texture.h
@@ -8,6 +8,6 @@ struct Texture {
         if (pixel) delete[] pixel;
     }
 
-    uint32_t* pixel{ nullptr };
+    float* pixel{ nullptr };
     glm::ivec2 resolution{ -1 };
 };
